# back/swagger.yaml
swagger: '2.0'
info:
  version: 1.0.0
  title: GoogleCalendar-API-docs
  description: GoogleCalendar API문서
  servers:
    - description: GoogleCalendar-Clone API
      url: 'http://158.247.214.79/'
    - description: GoogleCalendar-Clone dev
      url: 'http://localhost:80/'
...
paths:
...
  /api/privateEvent/createPrivateEvent:
    post:
      tags:
        - privateEvent
      summary: 개인 이벤트 만들기
      parameters:
        - name: Authorization
          in: header
          required: true
        - name: body
          in: body
          required: true
          schema:
            type: object
            properties:
              eventId:
                type: integer
                example: 1
              eventName:
                type: string
              color:
                type: integer
                example: '#0e0e0e'
              busy:
                type: integer
              memo:
                type: string
              startTime:
                type: string
                example: '2022-06-02T00:00:00'
              endTime:
                type: string
                example: '2022-06-02T00:00:00'
              allDay:
                type: integer
                example: 1
              alerts:
                type: array
                items:
                  type: object
                  properties:
                    type:
                      type: string
                      example: minute / hour / day / week
                    time:
                      type: integer
      responses:
        '200':
          $ref: '#/returnPrivateEvent'
        '500':
          $ref: '#/serverError'
  /api/privateEvent/editPrivateEvent:
    post:
      tags:
        - privateEvent
      summary: 개인 이벤트 수정
      parameters:
        - name: Authorization
          in: header
          required: true
        - name: body
          in: body
          required: true
          schema:
            type: object
            properties:
              eventId:
                type: integer
                example: 1
              eventName:
                type: string
              color:
                type: integer
                example: '#0e0e0e'
              busy:
                type: integer
              memo:
                type: string
              startTime:
                type: string
                example: '2022-06-02T00:00:00'
              endTime:
                type: string
                example: '2022-06-02T00:00:00'
              allDay:
                type: integer
                example: 1
              alerts:
                type: array
                items:
                  type: object
                  properties:
                    type:
                      type: string
                      example: minute / hour / day / week
                    time:
                      type: integer
      responses:
        '200':
          $ref: '#/returnPrivateEvent'
        '400':
          description: 수정할 개인 이벤트 조회 결과가 없을 경우
          schema:
            type: object
            properties:
              message:
                type: string
                example: 수정할 개인이벤트의 조회 결과가 없습니다
        '500':
          $ref: '#/serverError'
  /api/privateEvent/editPrivateEventColor:
    post:
      tags:
        - privateEvent
      summary: 개인 이벤트 색상 수정
      parameters:
        - name: Authorization
          in: header
          required: true
        - name: body
          in: body
          required: true
          schema:
            type: object
            properties:
              eventId:
                type: integer
              color:
                type: string
      responses:
        '200':
          $ref: '#/returnPrivateEvent'
        '400':
          description: 수정할 개인 이벤트 조회 결과가 없을 경우
          schema:
            type: object
            properties:
              message:
                type: string
                example: 수정할 개인이벤트의 조회 결과가 없습니다
        '500':
          $ref: '#/serverError'
  /api/privateEvent/deletePrivateEvent:
    post:
      tags:
        - privateEvent
      summary: 개인 이벤트 삭제
      parameters:
        - name: Authorization
          in: header
          required: true
        - name: body
          in: body
          required: true
          schema:
            type: object
            properties:
              eventId:
                type: integer
      responses:
        '200':
          $ref: '#/success'
        '500':
          $ref: '#/serverError'
...
  /api/alert/getAlerts:
    post:
      tags:
        - alert
      summary: 알림 가져오기
      parameters:
        - name: Authorization
          in: header
          required: true
      responses:
        '200':
          description: 성공
          schema:
            type: array
            items:
              type: object
              properties:
                id:
                  type: integer
                type:
                  type: string
                content:
                  type: string
                checked:
                  type: boolean
                calendarId:
                  type: integer
                eventDate:
                  type: string
        '500':
          $ref: '#/serverError'
  /api/alert/read:
    post:
      tags:
        - alert
      summary: 알림 읽음 표시하기
      parameters:
        - name: Authorization
          in: header
          required: true
        - name: body
          in: body
          required: true
          schema:
            type: object
            properties:
              alertId:
                type: integer
      responses:
        '200':
          $ref: '#/success'
        '500':
          $ref: '#/serverError'
...